Fix a bug in DateFormat, clean up some SimpleDateFormat tests.
We failed a couple of tests because of an assumption about the default
time zone:

  junit.framework.AssertionFailedError: Wrong format: " d" expected:  12 result:  11
  at java.text.SimpleDateFormatTest.test(SimpleDateFormatTest.java:55)
  at org.apache.harmony.tests.java.text.SimpleDateFormatTest.test_formatLjava_util_DateLjava_lang_StringBufferLjava_text_FieldPosition(SimpleDateFormatTest.java:347)

  junit.framework.AssertionFailedError: Wrong result: h input:  12 expected: Thu Jan 01 00:00:00 UTC 1970 result: Thu Jan 01 08:00:00 UTC 1970
  at java.text.SimpleDateFormatTest.parse(SimpleDateFormatTest.java:86)
  at org.apache.harmony.tests.java.text.SimpleDateFormatTest.test_parseLjava_lang_StringLjava_text_ParsePosition(SimpleDateFormatTest.java:546)

Cleaning up the tests and merging OldSimpleDateFormatTest into
SimpleDateFormatTest accidentally made one of our tests more
strict, at which point it failed because we weren't updating the
FieldPosition for 'z' or 'Z' when formatting or parsing. So this
patch fixes that by mapping the DateFormat.Field to the appropriate
int for Calendar.

Change-Id: If4b2f2c02d8ccd50d96d77df3b60e1c1d65ae666